Add a default region for `ec2 describe-regions`
In our case, we do not use the `default` AWS profile (so as to avoid using the wrong one). *Even when specifying a profile with a region set in the config*, the `ec2 describe-regions` command stumbles, with the error: `You must specify a region. You can also configure your region by running "aws configure"`. In this case, _any region_ can be used for this command, and it _correctly_ returns all the regions regardless of which region is specified. This change hard-codes a default.